requirements.txt
wheel>=0.22
Flask>=2.3.2
Flask-Restless-NG==3.0.0
flask-swagger-ui==4.11.1
PyYAML>=6.0
